+++ sdk/resourcemanager/compute/armcompute/models_serde.go
@@ -4528,88 +4528,6 @@ func (g *GalleryApplication) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type GalleryApplicationCustomAction.
-func (g GalleryApplicationCustomAction) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "description", g.Description)
-	populate(objectMap, "name", g.Name)
-	populate(objectMap, "parameters", g.Parameters)
-	populate(objectMap, "script", g.Script)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type GalleryApplicationCustomAction.
-func (g *GalleryApplicationCustomAction) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", g, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "description":
-			err = unpopulate(val, "Description", &g.Description)
-			delete(rawMsg, key)
-		case "name":
-			err = unpopulate(val, "Name", &g.Name)
-			delete(rawMsg, key)
-		case "parameters":
-			err = unpopulate(val, "Parameters", &g.Parameters)
-			delete(rawMsg, key)
-		case "script":
-			err = unpopulate(val, "Script", &g.Script)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", g, err)
-		}
-	}
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type GalleryApplicationCustomActionParameter.
-func (g GalleryApplicationCustomActionParameter) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "defaultValue", g.DefaultValue)
-	populate(objectMap, "description", g.Description)
-	populate(objectMap, "name", g.Name)
-	populate(objectMap, "required", g.Required)
-	populate(objectMap, "type", g.Type)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type GalleryApplicationCustomActionParameter.
-func (g *GalleryApplicationCustomActionParameter) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", g, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "defaultValue":
-			err = unpopulate(val, "DefaultValue", &g.DefaultValue)
-			delete(rawMsg, key)
-		case "description":
-			err = unpopulate(val, "Description", &g.Description)
-			delete(rawMsg, key)
-		case "name":
-			err = unpopulate(val, "Name", &g.Name)
-			delete(rawMsg, key)
-		case "required":
-			err = unpopulate(val, "Required", &g.Required)
-			delete(rawMsg, key)
-		case "type":
-			err = unpopulate(val, "Type", &g.Type)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", g, err)
-		}
-	}
-	return nil
-}
-
 // MarshalJSON implements the json.Marshaller interface for type GalleryApplicationList.
 func (g GalleryApplicationList) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
@@ -4644,7 +4562,6 @@ func (g *GalleryApplicationList) UnmarshalJSON(data []byte) error {
 // MarshalJSON implements the json.Marshaller interface for type GalleryApplicationProperties.
 func (g GalleryApplicationProperties) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
-	populate(objectMap, "customActions", g.CustomActions)
 	populate(objectMap, "description", g.Description)
 	populateTimeRFC3339(objectMap, "endOfLifeDate", g.EndOfLifeDate)
 	populate(objectMap, "eula", g.Eula)
@@ -4663,9 +4580,6 @@ func (g *GalleryApplicationProperties) UnmarshalJSON(data []byte) error {
 	for key, val := range rawMsg {
 		var err error
 		switch key {
-		case "customActions":
-			err = unpopulate(val, "CustomActions", &g.CustomActions)
-			delete(rawMsg, key)
 		case "description":
 			err = unpopulate(val, "Description", &g.Description)
 			delete(rawMsg, key)
@@ -4819,7 +4733,6 @@ func (g GalleryApplicationVersionProperties) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "provisioningState", g.ProvisioningState)
 	populate(objectMap, "publishingProfile", g.PublishingProfile)
 	populate(objectMap, "replicationStatus", g.ReplicationStatus)
-	populate(objectMap, "safetyProfile", g.SafetyProfile)
 	return json.Marshal(objectMap)
 }
 
@@ -4841,9 +4754,6 @@ func (g *GalleryApplicationVersionProperties) UnmarshalJSON(data []byte) error {
 		case "replicationStatus":
 			err = unpopulate(val, "ReplicationStatus", &g.ReplicationStatus)
 			delete(rawMsg, key)
-		case "safetyProfile":
-			err = unpopulate(val, "SafetyProfile", &g.SafetyProfile)
-			delete(rawMsg, key)
 		}
 		if err != nil {
 			return fmt.Errorf("unmarshalling type %T: %v", g, err)
@@ -4856,7 +4766,6 @@ func (g *GalleryApplicationVersionProperties) UnmarshalJSON(data []byte) error {
 func (g GalleryApplicationVersionPublishingProfile) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
 	populate(objectMap, "advancedSettings", g.AdvancedSettings)
-	populate(objectMap, "customActions", g.CustomActions)
 	populate(objectMap, "enableHealthCheck", g.EnableHealthCheck)
 	populateTimeRFC3339(objectMap, "endOfLifeDate", g.EndOfLifeDate)
 	populate(objectMap, "excludeFromLatest", g.ExcludeFromLatest)
@@ -4883,9 +4792,6 @@ func (g *GalleryApplicationVersionPublishingProfile) UnmarshalJSON(data []byte) 
 		switch key {
 		case "advancedSettings":
 			err = unpopulate(val, "AdvancedSettings", &g.AdvancedSettings)
-			delete(rawMsg, key)
-		case "customActions":
-			err = unpopulate(val, "CustomActions", &g.CustomActions)
 			delete(rawMsg, key)
 		case "enableHealthCheck":
 			err = unpopulate(val, "EnableHealthCheck", &g.EnableHealthCheck)
@@ -4922,33 +4828,6 @@ func (g *GalleryApplicationVersionPublishingProfile) UnmarshalJSON(data []byte) 
 			delete(rawMsg, key)
 		case "targetRegions":
 			err = unpopulate(val, "TargetRegions", &g.TargetRegions)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", g, err)
-		}
-	}
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type GalleryApplicationVersionSafetyProfile.
-func (g GalleryApplicationVersionSafetyProfile) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "allowDeletionOfReplicatedLocations", g.AllowDeletionOfReplicatedLocations)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type GalleryApplicationVersionSafetyProfile.
-func (g *GalleryApplicationVersionSafetyProfile) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", g, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "allowDeletionOfReplicatedLocations":
-			err = unpopulate(val, "AllowDeletionOfReplicatedLocations", &g.AllowDeletionOfReplicatedLocations)
 			delete(rawMsg, key)
 		}
 		if err != nil {
@@ -5056,33 +4935,6 @@ func (g *GalleryArtifactPublishingProfileBase) UnmarshalJSON(data []byte) error 
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type GalleryArtifactSafetyProfileBase.
-func (g GalleryArtifactSafetyProfileBase) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "allowDeletionOfReplicatedLocations", g.AllowDeletionOfReplicatedLocations)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type GalleryArtifactSafetyProfileBase.
-func (g *GalleryArtifactSafetyProfileBase) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", g, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "allowDeletionOfReplicatedLocations":
-			err = unpopulate(val, "AllowDeletionOfReplicatedLocations", &g.AllowDeletionOfReplicatedLocations)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", g, err)
-		}
-	}
-	return nil
-}
-
 // MarshalJSON implements the json.Marshaller interface for type GalleryArtifactSource.
 func (g GalleryArtifactSource) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
@@ -5110,41 +4962,11 @@ func (g *GalleryArtifactSource) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type GalleryArtifactVersionFullSource.
-func (g GalleryArtifactVersionFullSource) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "communityGalleryImageId", g.CommunityGalleryImageID)
-	populate(objectMap, "id", g.ID)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type GalleryArtifactVersionFullSource.
-func (g *GalleryArtifactVersionFullSource) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", g, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "communityGalleryImageId":
-			err = unpopulate(val, "CommunityGalleryImageID", &g.CommunityGalleryImageID)
-			delete(rawMsg, key)
-		case "id":
-			err = unpopulate(val, "ID", &g.ID)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", g, err)
-		}
-	}
-	return nil
-}
-
 // MarshalJSON implements the json.Marshaller interface for type GalleryArtifactVersionSource.
 func (g GalleryArtifactVersionSource) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
 	populate(objectMap, "id", g.ID)
+	populate(objectMap, "uri", g.URI)
 	return json.Marshal(objectMap)
 }
 
@@ -5159,6 +4981,9 @@ func (g *GalleryArtifactVersionSource) UnmarshalJSON(data []byte) error {
 		switch key {
 		case "id":
 			err = unpopulate(val, "ID", &g.ID)
+			delete(rawMsg, key)
+		case "uri":
+			err = unpopulate(val, "URI", &g.URI)
 			delete(rawMsg, key)
 		}
 		if err != nil {
@@ -5233,41 +5058,6 @@ func (g *GalleryDiskImage) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "source":
 			err = unpopulate(val, "Source", &g.Source)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", g, err)
-		}
-	}
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type GalleryDiskImageSource.
-func (g GalleryDiskImageSource) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "id", g.ID)
-	populate(objectMap, "storageAccountId", g.StorageAccountID)
-	populate(objectMap, "uri", g.URI)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type GalleryDiskImageSource.
-func (g *GalleryDiskImageSource) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", g, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "id":
-			err = unpopulate(val, "ID", &g.ID)
-			delete(rawMsg, key)
-		case "storageAccountId":
-			err = unpopulate(val, "StorageAccountID", &g.StorageAccountID)
-			delete(rawMsg, key)
-		case "uri":
-			err = unpopulate(val, "URI", &g.URI)
 			delete(rawMsg, key)
 		}
 		if err != nil {
@@ -5689,7 +5479,6 @@ func (g GalleryImageVersionProperties) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "provisioningState", g.ProvisioningState)
 	populate(objectMap, "publishingProfile", g.PublishingProfile)
 	populate(objectMap, "replicationStatus", g.ReplicationStatus)
-	populate(objectMap, "safetyProfile", g.SafetyProfile)
 	populate(objectMap, "storageProfile", g.StorageProfile)
 	return json.Marshal(objectMap)
 }
@@ -5711,9 +5500,6 @@ func (g *GalleryImageVersionProperties) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "replicationStatus":
 			err = unpopulate(val, "ReplicationStatus", &g.ReplicationStatus)
-			delete(rawMsg, key)
-		case "safetyProfile":
-			err = unpopulate(val, "SafetyProfile", &g.SafetyProfile)
 			delete(rawMsg, key)
 		case "storageProfile":
 			err = unpopulate(val, "StorageProfile", &g.StorageProfile)
@@ -5772,41 +5558,6 @@ func (g *GalleryImageVersionPublishingProfile) UnmarshalJSON(data []byte) error 
 			delete(rawMsg, key)
 		case "targetRegions":
 			err = unpopulate(val, "TargetRegions", &g.TargetRegions)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", g, err)
-		}
-	}
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type GalleryImageVersionSafetyProfile.
-func (g GalleryImageVersionSafetyProfile) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "allowDeletionOfReplicatedLocations", g.AllowDeletionOfReplicatedLocations)
-	populate(objectMap, "policyViolations", g.PolicyViolations)
-	populate(objectMap, "reportedForPolicyViolation", g.ReportedForPolicyViolation)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type GalleryImageVersionSafetyProfile.
-func (g *GalleryImageVersionSafetyProfile) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", g, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "allowDeletionOfReplicatedLocations":
-			err = unpopulate(val, "AllowDeletionOfReplicatedLocations", &g.AllowDeletionOfReplicatedLocations)
-			delete(rawMsg, key)
-		case "policyViolations":
-			err = unpopulate(val, "PolicyViolations", &g.PolicyViolations)
-			delete(rawMsg, key)
-		case "reportedForPolicyViolation":
-			err = unpopulate(val, "ReportedForPolicyViolation", &g.ReportedForPolicyViolation)
 			delete(rawMsg, key)
 		}
 		if err != nil {
@@ -6985,37 +6736,6 @@ func (l *LastPatchInstallationSummary) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "status":
 			err = unpopulate(val, "Status", &l.Status)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", l, err)
-		}
-	}
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type LatestGalleryImageVersion.
-func (l LatestGalleryImageVersion) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "latestVersionName", l.LatestVersionName)
-	populate(objectMap, "location", l.Location)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type LatestGalleryImageVersion.
-func (l *LatestGalleryImageVersion) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", l, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "latestVersionName":
-			err = unpopulate(val, "LatestVersionName", &l.LatestVersionName)
-			delete(rawMsg, key)
-		case "location":
-			err = unpopulate(val, "Location", &l.Location)
 			delete(rawMsg, key)
 		}
 		if err != nil {
@@ -8523,37 +8243,6 @@ func (p *Plan) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "publisher":
 			err = unpopulate(val, "Publisher", &p.Publisher)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", p, err)
-		}
-	}
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type PolicyViolation.
-func (p PolicyViolation) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "category", p.Category)
-	populate(objectMap, "details", p.Details)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type PolicyViolation.
-func (p *PolicyViolation) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", p, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "category":
-			err = unpopulate(val, "Category", &p.Category)
-			delete(rawMsg, key)
-		case "details":
-			err = unpopulate(val, "Details", &p.Details)
 			delete(rawMsg, key)
 		}
 		if err != nil {
@@ -11792,13 +11481,11 @@ func (s SharedGalleryImageProperties) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "architecture", s.Architecture)
 	populate(objectMap, "disallowed", s.Disallowed)
 	populateTimeRFC3339(objectMap, "endOfLifeDate", s.EndOfLifeDate)
-	populate(objectMap, "eula", s.Eula)
 	populate(objectMap, "features", s.Features)
 	populate(objectMap, "hyperVGeneration", s.HyperVGeneration)
 	populate(objectMap, "identifier", s.Identifier)
 	populate(objectMap, "osState", s.OSState)
 	populate(objectMap, "osType", s.OSType)
-	populate(objectMap, "privacyStatementUri", s.PrivacyStatementURI)
 	populate(objectMap, "purchasePlan", s.PurchasePlan)
 	populate(objectMap, "recommended", s.Recommended)
 	return json.Marshal(objectMap)
@@ -11822,9 +11509,6 @@ func (s *SharedGalleryImageProperties) UnmarshalJSON(data []byte) error {
 		case "endOfLifeDate":
 			err = unpopulateTimeRFC3339(val, "EndOfLifeDate", &s.EndOfLifeDate)
 			delete(rawMsg, key)
-		case "eula":
-			err = unpopulate(val, "Eula", &s.Eula)
-			delete(rawMsg, key)
 		case "features":
 			err = unpopulate(val, "Features", &s.Features)
 			delete(rawMsg, key)
@@ -11839,9 +11523,6 @@ func (s *SharedGalleryImageProperties) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "osType":
 			err = unpopulate(val, "OSType", &s.OSType)
-			delete(rawMsg, key)
-		case "privacyStatementUri":
-			err = unpopulate(val, "PrivacyStatementURI", &s.PrivacyStatementURI)
 			delete(rawMsg, key)
 		case "purchasePlan":
 			err = unpopulate(val, "PurchasePlan", &s.PurchasePlan)
@@ -12831,7 +12512,6 @@ func (s *SystemData) UnmarshalJSON(data []byte) error {
 func (t TargetRegion) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
 	populate(objectMap, "encryption", t.Encryption)
-	populate(objectMap, "excludeFromLatest", t.ExcludeFromLatest)
 	populate(objectMap, "name", t.Name)
 	populate(objectMap, "regionalReplicaCount", t.RegionalReplicaCount)
 	populate(objectMap, "storageAccountType", t.StorageAccountType)
@@ -12849,9 +12529,6 @@ func (t *TargetRegion) UnmarshalJSON(data []byte) error {
 		switch key {
 		case "encryption":
 			err = unpopulate(val, "Encryption", &t.Encryption)
-			delete(rawMsg, key)
-		case "excludeFromLatest":
-			err = unpopulate(val, "ExcludeFromLatest", &t.ExcludeFromLatest)
 			delete(rawMsg, key)
 		case "name":
 			err = unpopulate(val, "Name", &t.Name)